+++ imessage-exporter/src/exporters/mod.rs
@@ -1,4 +1,4 @@
-pub mod exporter;
+pub mod formatter;
 pub mod html;
 pub mod shared;
 pub mod txt;
